Gameboy: Move away from Singleton of Cartridge, CPU, and MMU
The Singleton design was making it extremely difficult to unit test.
A better alternative is to create each of these as dependencies on the
main Gameboy class and pass the reference when initializing. This allows
for creating mock instances of each class that can be more easily
targeted for unit testing.