Cartridge: Handle the file loading outside of the Cartridge class
Previously the Cartridge class would handle the logic of ensuring a file
exists on the system, and loading it into the rom_data. It is better to
separate this functionality outside of the class and implement it as a
file utility.

Additionally, this patch changes the m_data type from a char array to
std::vector.