Register: Overload the set method for FlagRegister
As the bottom half of the F register should not be used, this patch
overloads the set method for the FlagRegister so that it will ensure
that the bottom nibble does not get assigned a value.

Additionally, it is better to call the set methods in the WholeRegister
so that the overloaded F set will get called whenever there is a write
to the AF register.

This solves the POP AF fail when conducting Blargg's 01-special.gb test