Gameboy: Add support for wasm rendering
This provides the basis for the wasm port. This current patch will
compile with the emscripten toolchain, but does not render anything as
the graphics have still yet to be implemented.

Further refactoring may be necessary to account for the different
platform types (i.e. how to handle file/rom importing)